修改tapport字段
**复现步骤**

**原因和解决方案**

* 添加字段区分具体的内容

**影响范围**

**验证方案**

**涉及分支**

* 600
* master

**检查项**

- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 编译通过
- [ ] 单元测试通过

/assign @xiangyang

/cc @yuanchao @jinjie @zhuofeng @wangtuo

+++ zerodoc/tag_test.go
@@ -368,7 +368,7 @@ func TestDirectionToTAPSide(t *testing.T) {
 func TestPutTAPPort(t *testing.T) {
 	bs := make([]byte, 16)
 	for _, tc := range []struct {
-		input  uint32
+		input  uint64
 		output string
 	}{
 		{0, "00000000"},
